Changed logo and simplified endpoints
Now shows top, upcoming,  new, and popular films. Additionally showcases actors starring in the film, and provides detailed information on them. Continuations are a hidden feature, that essentially allows the user to ask which actors are in a given film, and which films contain a given actor. This is all do using the IMDB API